03-multiagent-demo/requirements.txt
strands-agents>=1.27.0
boto3>=1.35.0
